+++ packages/react/Chip/src/components/Skeleton/skeleton.types.ts
@@ -1,0 +1,6 @@
+import { SkeletonProps as SkeletonBaseProps } from "@nimbus-ds/skeleton";
+
+export type SkeletonProps = Partial<Pick<SkeletonBaseProps, "width">> &
+  Partial<Pick<SkeletonBaseProps, "height">> & {
+    "data-testid"?: string;
+  };