+++ packages/react/Text/src/components/Skeleton/skeleton.spec.tsx
@@ -13,25 +13,37 @@ describe("GIVEN <Text.Skeleton />", () => {
     it('THEN should render skeleton size="small"', () => {
       makeSut({ size: "small" });
       const skeleton = screen.getByTestId("skeleton-element");
-      expect(skeleton.getAttribute("style")).toContain("2rem");
-      expect(skeleton.getAttribute("style")).toContain("1rem");
-      expect(skeleton.getAttribute("style")).toContain("0.25rem");
+      expect(skeleton.getAttribute("style")).toMatch(/--width__\w{0,9}: 2rem;/);
+      expect(skeleton.getAttribute("style")).toMatch(
+        /--height__\w{0,9}: 1rem;/
+      );
+      expect(skeleton.getAttribute("style")).toMatch(
+        /--borderRadius__\w{0,9}: 0.25rem;/
+      );
     });
 
     it('THEN should render skeleton size="caption"', () => {
       makeSut({ size: "caption" });
       const skeleton = screen.getByTestId("skeleton-element");
-      expect(skeleton.getAttribute("style")).toContain("2rem");
-      expect(skeleton.getAttribute("style")).toContain("1.25rem");
-      expect(skeleton.getAttribute("style")).toContain("0.25rem");
+      expect(skeleton.getAttribute("style")).toMatch(/--width__\w{0,9}: 2rem;/);
+      expect(skeleton.getAttribute("style")).toMatch(
+        /--height__\w{0,9}: 1.25rem;/
+      );
+      expect(skeleton.getAttribute("style")).toMatch(
+        /--borderRadius__\w{0,9}: 0.25rem;/
+      );
     });
 
     it('THEN should render skeleton type="base"', () => {
       makeSut({ size: "base" });
       const skeleton = screen.getByTestId("skeleton-element");
-      expect(skeleton.getAttribute("style")).toContain("2rem");
-      expect(skeleton.getAttribute("style")).toContain("1.125rem");
-      expect(skeleton.getAttribute("style")).toContain("0.25rem");
+      expect(skeleton.getAttribute("style")).toMatch(/--width__\w{0,9}: 2rem;/);
+      expect(skeleton.getAttribute("style")).toMatch(
+        /--height__\w{0,9}: 1.125rem;/
+      );
+      expect(skeleton.getAttribute("style")).toMatch(
+        /--borderRadius__\w{0,9}: 0.25rem;/
+      );
     });
   });
 });