+++ packages/react/Text/src/text.stories.tsx
@@ -7,12 +7,19 @@ import { Text } from "./Text";
 export default {
   title: "Atomic/Text",
   component: Text,
+  subcomponents: { "Text.Skeleton": Text.Skeleton },
+  argTypes: {
+    children: { control: { disable: true } },
+  },
   parameters: {
     withA11y: { decorators: [withA11y] },
   },
 } as ComponentMeta<typeof Text>;
 
 const Template: ComponentStory<typeof Text> = (args) => <Text {...args} />;
+const SkeletonTemplate: ComponentStory<typeof Text.Skeleton> = (args) => (
+  <Text.Skeleton {...args} />
+);
 
 export const p = Template.bind({});
 p.args = {
@@ -28,30 +35,33 @@ span.args = {
 
 export const primary = Template.bind({});
 primary.args = {
-  appearance: "primary",
+  color: "primary.textLow",
   children: "Lorem ipsum dolor sit amet",
 };
 
 export const success = Template.bind({});
 success.args = {
-  appearance: "success",
+  color: "success.textLow",
   children: "Lorem ipsum dolor sit amet",
 };
 
 export const warning = Template.bind({});
 warning.args = {
-  appearance: "warning",
+  color: "warning.textLow",
   children: "Lorem ipsum dolor sit amet",
 };
 
 export const danger = Template.bind({});
 danger.args = {
-  appearance: "danger",
+  color: "danger.textLow",
   children: "Lorem ipsum dolor sit amet",
 };
 
 export const neutral = Template.bind({});
 neutral.args = {
-  appearance: "neutral",
+  color: "neutral.textLow",
   children: "Lorem ipsum dolor sit amet",
 };
+
+export const skeleton = SkeletonTemplate.bind({});
+skeleton.args = {};