+++ packages/react/Title/src/components/Skeleton/skeleton.stories.tsx
@@ -5,7 +5,7 @@ import { withA11y } from "@storybook/addon-a11y";
 import { Skeleton } from "./Skeleton";
 
 export default {
-  title: "Atomic/Title/Title.Skeleton",
+  title: "Skeleton/Title",
   component: Skeleton,
   parameters: {
     withA11y: { decorators: [withA11y] },
@@ -16,32 +16,5 @@ const Template: ComponentStory<typeof Skeleton> = (args) => (
   <Skeleton {...args} />
 );
 
-export const h1 = Template.bind({});
-h1.args = {
-  as: "h1",
-};
-
-export const h2 = Template.bind({});
-h2.args = {
-  as: "h2",
-};
-
-export const h3 = Template.bind({});
-h3.args = {
-  as: "h3",
-};
-
-export const h4 = Template.bind({});
-h4.args = {
-  as: "h4",
-};
-
-export const h5 = Template.bind({});
-h5.args = {
-  as: "h5",
-};
-
-export const h6 = Template.bind({});
-h6.args = {
-  as: "h6",
-};
+export const title = Template.bind({});
+title.args = {};