+++ packages/react/Title/webpack.config.js
@@ -1,5 +1,4 @@
 const path = require("path");
-const { VanillaExtractPlugin } = require("@vanilla-extract/webpack-plugin");
 const TerserJSPlugin = require("terser-webpack-plugin");
 
 module.exports = {
@@ -17,21 +16,20 @@ module.exports = {
   module: {
     rules: [
       {
+        test: /\.css$/,
+        use: ["style-loader", "css-loader"],
+      },
+      {
         test: /\.tsx?$/,
         loader: "ts-loader",
         exclude: /node_modules/,
       },
-      {
-        test: /\.css$/i,
-        use: ["style-loader", "css-loader"],
-      },
     ],
   },
-  plugins: [new VanillaExtractPlugin()],
   resolve: {
     alias: {
       "@nimbus-ds/skeleton": path.resolve(__dirname, "../Skeleton/src"),
-      "@nimbus-ds/styles": path.resolve(__dirname, "../../styles/src"),
+      "@nimbus-ds/styles": path.resolve(__dirname, "../../styles/dist"),
     },
     extensions: [".tsx", ".ts", ".js"],
   },
@@ -40,6 +38,8 @@ module.exports = {
     minimizer: [new TerserJSPlugin()],
   },
   externals: {
+    "@nimbus-ds/styles": "@nimbus-ds/styles",
+    "@nimbus-ds/skeleton": "@nimbus-ds/skeleton",
     react: "react",
     "react-dom": "react-dom",
   },