+++ packages/styles/src/packages/tag/index.ts
@@ -1,0 +1,5 @@
+import { style } from "./tag.style.css";
+
+export const tag = {
+  style,
+};