+++ packages/styles/src/packages/text/index.ts
@@ -1,0 +1,8 @@
+import { style } from "./text.style.css";
+import { sprinkle, properties } from "./text.sprinkle.css";
+
+export const text = {
+  style,
+  sprinkle,
+  properties,
+};